Move all jobs into dedicated modules
